refactor: Create CategorySelectionTemplate for rendering category options in select element
