+++ src/templates/CategorySelectionTemplate.ts
